refactor: migrate application structure to MVC pattern
- Split monolithic app.py into models, controllers, and services
- Implement Application Factory pattern in app/__init__.py
- Create PresentationModel for data persistence logic
- Move configuration to config.py
- Update unit and integration tests to support new structure
- Clean up unused files and add .env.example
