+++ .env.example
@@ -1,0 +1,9 @@
+# Google Gemini API Key
+# Get your API key here: https://aistudio.google.com/app/apikey
+GEMINI_API_KEY=your_gemini_api_key_here
+
+# Flask Configuration
+FLASK_APP=run.py
+FLASK_ENV=development
+# Secret key for session management (optional but recommended)
+SECRET_KEY=your_secret_key_here
